Move "refreshConfig" to config.ts
